cleaned up config documentation details, added file headers
