+++ isofit/configs/base_config.py
@@ -1,4 +1,22 @@
-import re
+#! /usr/bin/env python3
+#
+#  Copyright 2018 California Institute of Technology
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+# ISOFIT: Imaging Spectrometer Optimal FITting
+# Author: Philip G. Brodrick, philip.brodrick@jpl.nasa.gov
+
 from collections import OrderedDict
 from typing import Dict, List, Type
 import numpy as np
@@ -8,34 +26,6 @@ class BaseConfigSection(object):
     """
     Base Configuration Section from which all Configuration Sections inherit. Handles shared functionality like getting,
     setting, and cleaning configuration options.
-
-    How to use sections:  Configs and ConfigSections are the tools we use to handle the numerous parameters
-    associated with atmospheric correction. In general, we want to validate that ConfigSections have options with
-    the expected values and relationships, and we use two important methods to do this without duplicating too much code.
-
-    To validate that attributes have the correct type, ensure that the attributes are on the config section and have an
-    associated hidden attribute with a particular name pattern. Specifically, given an attribute named 'attribute', the
-    hidden attribute should be named '_attribute_type' and its value should be the type expected for that attribute.
-    Methods on the BaseConfigSection will ensure that this attribute type is checked and errors will be raised to the user
-    if it's not appropriate. Example:
-
-    ```
-    class GenericConfigSection(BaseConfigSection):
-        _attribute_type = list                  <-- Used to validate attributes have correct types
-        attribute = DEFAULT_REQUIRED_VALUE
-    ```
-
-    To validate that attributes have appropriate relationships or characteristics, use the hidden _check_config_validity
-    method to add more detailed validation checks. Simply return a list of string descriptions of errors from the
-    method as demonstrated:
-
-    ```
-    def _check_config_validity(self) -> List[str]:
-        errors = list()
-        if self.attribute_min >= self.attribute_max:
-            errors.append('attribute_min must be less than attribute_max.')
-        return errors
-    ```
     """
 
     def __init__(self) -> None:
@@ -177,7 +167,7 @@ class BaseConfigSection(object):
         return elements[element_names.index(name)]
 
 
-def snake_to_camel(word: str) -> None:
+def snake_to_camel(word: str) -> str:
     """ Function to convert snake case to camel case, e.g.
     snake_to_camel -> SnakeToCamel
     Args:
