+++ isofit/configs/sections/output_config.py
@@ -1,3 +1,21 @@
+#! /usr/bin/env python3
+#
+#  Copyright 2018 California Institute of Technology
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+# ISOFIT: Imaging Spectrometer Optimal FITting
+# Author: Philip G. Brodrick, philip.brodrick@jpl.nasa.gov
 
 from typing import Dict, List, Type
 from isofit.configs.base_config import BaseConfigSection
